refactor: tests: Move MessageBox into test_boxes.
The message box has been moved to test_boxes to have all tests for
boxes in the same place. The fixture for `compose_box_is_open` which
is now needed for both files is moved to the Controller Fixtures
section of `conftest.py`.
